+++ lib/Component/Exception/ReviewException.php
@@ -1,0 +1,7 @@
+<?php namespace Vankosoft\PaymentBundle\Component\Exception;
+
+use RuntimeException;
+
+class ReviewException extends RuntimeException
+{
+}