+++ lib/Model/Interfaces/CommenterInterface.php
@@ -1,0 +1,8 @@
+<?php namespace Vankosoft\PaymentBundle\Model\Interfaces;
+
+use Doctrine\Common\Collections\Collection;
+
+interface CommenterInterface
+{
+    public function getComments(): Collection;
+}