+++ lib/Model/Review.php
@@ -1,0 +1,9 @@
+<?php namespace Vankosoft\PaymentBundle\Model;
+
+use Sylius\Component\Review\Model\Review as BaseReview;
+use Vankosoft\PaymentBundle\Model\Interfaces\ReviewInterface;
+
+class Review extends BaseReview implements ReviewInterface
+{
+    
+}